LayoutTests/imported/w3c/web-platform-tests/css/motion/animation/offset-anchor-composition-expected.txt


FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 40px] to add [160px 140px] at (-0.25) should be [75px 75px] assert_equals: expected "75px 75px " but got "35px 15px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 40px] to add [160px 140px] at (0) should be [100px 100px] assert_equals: expected "100px 100px " but got "60px 40px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 40px] to add [160px 140px] at (0.25) should be [125px 125px] assert_equals: expected "125px 125px " but got "85px 65px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 40px] to add [160px 140px] at (0.5) should be [150px 150px] assert_equals: expected "150px 150px " but got "110px 90px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 40px] to add [160px 140px] at (0.75) should be [175px 175px] assert_equals: expected "175px 175px " but got "135px 115px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 40px] to add [160px 140px] at (1) should be [200px 200px] assert_equals: expected "200px 200px " but got "160px 140px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 40px] to add [160px 140px] at (1.25) should be [225px 225px] assert_equals: expected "225px 225px " but got "185px 165px "
FAIL Compositing: property <offset-anchor> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (-0.25) should be [110% 105%] assert_equals: expected "110 % 105 % " but got "70 % 85 % "
FAIL Compositing: property <offset-anchor> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (0) should be [100% 100%] assert_equals: expected "100 % 100 % " but got "60 % 80 % "
FAIL Compositing: property <offset-anchor> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (0.25) should be [90% 95%] assert_equals: expected "90 % 95 % " but got "50 % 75 % "
FAIL Compositing: property <offset-anchor> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (0.5) should be [80% 90%] assert_equals: expected "80 % 90 % " but got "40 % 70 % "
FAIL Compositing: property <offset-anchor> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (0.75) should be [70% 85%] assert_equals: expected "70 % 85 % " but got "30 % 65 % "
FAIL Compositing: property <offset-anchor> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (1) should be [60% 80%] assert_equals: expected "60 % 80 % " but got "20 % 60 % "
FAIL Compositing: property <offset-anchor> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (1.25) should be [50% 75%] assert_equals: expected "50 % 75 % " but got "10 % 55 % "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from replace [100px 200px] to add [160px 40px] at (-0.25) should be [75px 225px] assert_equals: expected "75px 225px " but got "85px 240px "
PASS Compositing: property <offset-anchor> underlying [40px 60px] from replace [100px 200px] to add [160px 40px] at (0) should be [100px 200px]
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from replace [100px 200px] to add [160px 40px] at (0.25) should be [125px 175px] assert_equals: expected "125px 175px " but got "115px 160px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from replace [100px 200px] to add [160px 40px] at (0.5) should be [150px 150px] assert_equals: expected "150px 150px " but got "130px 120px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from replace [100px 200px] to add [160px 40px] at (0.75) should be [175px 125px] assert_equals: expected "175px 125px " but got "145px 80px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from replace [100px 200px] to add [160px 40px] at (1) should be [200px 100px] assert_equals: expected "200px 100px " but got "160px 40px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from replace [100px 200px] to add [160px 40px] at (1.25) should be [225px 75px] assert_equals: expected "225px 75px " but got "175px 0px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 140px] to replace [200px 100px] at (-0.25) should be [75px 225px] assert_equals: expected "75px 225px " but got "25px 150px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 140px] to replace [200px 100px] at (0) should be [100px 200px] assert_equals: expected "100px 200px " but got "60px 140px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 140px] to replace [200px 100px] at (0.25) should be [125px 175px] assert_equals: expected "125px 175px " but got "95px 130px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 140px] to replace [200px 100px] at (0.5) should be [150px 150px] assert_equals: expected "150px 150px " but got "130px 120px "
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 140px] to replace [200px 100px] at (0.75) should be [175px 125px] assert_equals: expected "175px 125px " but got "165px 110px "
PASS Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 140px] to replace [200px 100px] at (1) should be [200px 100px]
FAIL Compositing: property <offset-anchor> underlying [40px 60px] from add [60px 140px] to replace [200px 100px] at (1.25) should be [225px 75px] assert_equals: expected "225px 75px " but got "235px 90px "
